HSEARCH-2924 Move JSR-352 SPI classes to properly named (*.spi) packages
