+++ jsr352/core/src/main/java/org/hibernate/search/jsr352/context/jpa/spi/EntityManagerFactoryRegistry.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.jsr352.context.jpa;
+package org.hibernate.search.jsr352.context.jpa.spi;
 
 import javax.persistence.EntityManagerFactory;
