+++ jsr352/core/src/main/java/org/hibernate/search/jsr352/massindexing/impl/steps/lucene/EntityReader.java
@@ -27,8 +27,8 @@ import org.hibernate.ScrollableResults;
 import org.hibernate.Session;
 import org.hibernate.criterion.Criterion;
 import org.hibernate.query.Query;
-import org.hibernate.search.jsr352.context.jpa.EntityManagerFactoryRegistry;
-import org.hibernate.search.jsr352.inject.scope.HibernateSearchPartitionScoped;
+import org.hibernate.search.jsr352.context.jpa.spi.EntityManagerFactoryRegistry;
+import org.hibernate.search.jsr352.inject.scope.spi.HibernateSearchPartitionScoped;
 import org.hibernate.search.jsr352.logging.impl.Log;
 import org.hibernate.search.jsr352.massindexing.MassIndexingJobParameters;
 import org.hibernate.search.jsr352.massindexing.MassIndexingJobParameters.Defaults;
