+++ jsr352/jberet/src/main/java/org/hibernate/search/jsr352/jberet/context/jpa/impl/CDIEntityManagerFactoryRegistry.java
@@ -21,7 +21,7 @@ import javax.inject.Singleton;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.PersistenceUnit;
 
-import org.hibernate.search.jsr352.context.jpa.EntityManagerFactoryRegistry;
+import org.hibernate.search.jsr352.context.jpa.spi.EntityManagerFactoryRegistry;
 import org.hibernate.search.jsr352.logging.impl.Log;
 import org.hibernate.search.util.logging.impl.LoggerFactory;
